[REF] web: convert model>sample_server tests to hoot
task 3705027
